Escape container attributes for every template
Messages go out with parse_mode HTML while container names, image tags
and custom labels were interpolated raw. The README encourages arbitrary
text in labels, and a single '<' from one is enough for Telegram to
reject the whole message — the notification is then simply lost.

The escaping happens once, on the attributes, before the template sees
them. Doing it inside templates.js would have covered the shipped
templates only; section 3 of the README tells people to mount their own,
and those would each have had to remember. This way they are safe
without changing anything.

The helper moves to its own module because telegram.js needs it too, for
the payload it puts in error notifications.

Verified with the custom label pattern the README documents: a label
containing "<b>bold</b> & <script>alert(1)</script>" now arrives escaped
instead of breaking the message.
